Refactor everything but cmd_handler.py
